refactor: #8 AppTextStyle 패키지 이동
