+++ designsystem/build.gradle.kts
@@ -19,7 +19,6 @@ kotlin {
 
     sourceSets {
         commonMain.dependencies {
-            implementation(project(":domain"))
             implementation(libs.bundles.compose.multiplatform)
         }
         commonTest.dependencies {
